Refactor: Next.js의 Script 컴포넌트를 활용하는 방식으로 변경한다
