+++ pages/_app.tsx
@@ -4,12 +4,31 @@ import { globalStyles, theme } from '@/styles';
 import { Layout } from '@/components';
 import { SessionProvider } from 'next-auth/react';
 import { useGoogleAnalytics } from '@/hooks';
+import Script from 'next/script';
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
   useGoogleAnalytics();
 
   return (
     <>
+      <Script
+        strategy="afterInteractive"
+        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.GOOGLE_ANALYTICS}`}
+      />
+      <Script
+        id="gtag-init"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', '${process.env.GOOGLE_ANALYTICS}', {
+              page_path: window.location.pathname,
+            });
+        `,
+        }}
+      />
       <Global styles={globalStyles} />
       <SessionProvider session={pageProps.session}>
         <ThemeProvider theme={theme}>
